Remove icons from buttons with Cleanlooks
The icon buttons in cleanlooks were modeled after an early version of
the Gtk 2.0 Clearlooks theme. At the time, GNOME was using
icons on all buttons. For the past couple of years, GNOME has dropped
icons on buttons by default and as this is the common behavior on
other platforms as well, I think it is time to remove them.

The button size hack was also removed due to previously emulating
a GNOME behavior. The end result at the moment is only that it makes
the button height inconsistent with other widgets.

Change-Id: I22fe8c0b3a799a807cfff74a81cebf50e8ec33d3
Reviewed-by: J-P Nurmi <j-p.nurmi@nokia.com>
Reviewed-by: Gabriel de Dietrich <gabriel.dietrich-de@nokia.com>