QWizard: use less stack space
There will hardly ever be 512 buttons in a wizard dialog, so specify
a somewhat saner preallocation number for the QVarLengthArray, e.g.
the number of available buttons.

Yes, there's also Stretch, but it's unlikely that _all_ buttons
plus stretch are used in one dialog, and QVarLengthArray will
fall back to the heap in that case, of course.

Change-Id: Ida09ab73340e9fb2aa80db37aeed8b8476ffe7fe
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>